% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{a_operator}
\alias{a_operator}
\title{The A operator is given by the integration of a set of differential equations
between two consecutive nodes. So, defined the set in the time interval
[t_{i-1}, t_i], where k species are present in the phylogeny, as:}
\usage{
a_operator(q_vector, transition_matrix, time_interval, methode = "lsoda",
  abstol = 1e-10, reltol = 1e-12, precision = 50L,
  atol_min = 1e-08, rtol_min = 1e-10, atol_max = 1e-20,
  rtol_max = 1e-88, atol_step = -2, rtol_step = -6)
}
\arguments{
\item{q_vector}{the \code{Q} vector from
'Etienne et al. 2012 - Proc. R. Soc. B'.}

\item{transition_matrix}{something}

\item{time_interval}{something}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
  \item \code{ode45}: use \code{ode45} and \code{deSolve::ode}
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{abstol}{absolute error tolerance for
the numerical integration using deSolve.}

\item{reltol}{relative error tolerance for
the numerical integration using deSolve.}

\item{precision}{something}

\item{atol_min}{min for atol}

\item{rtol_min}{min for rtol}

\item{atol_max}{max for atol}

\item{rtol_max}{max for rtol}

\item{atol_step}{step for atol}

\item{rtol_step}{step for rtol}
}
\description{
d
--Q^k_m(t) = SUM_n(M^k_m,n * Q^k_n(t)
dt
}
\details{
where m, n, label the amount of unseen species in the phylogeny,
A is thus defined as:

A(t_i - t_{i-1}) = exp(M(t_k - t_{k-1})
}
\author{
Giovanni Laudanno
}
