% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{alpha_analysis}
\alias{alpha_analysis}
\title{Internal MBD function}
\usage{
alpha_analysis(brts, pars, tips_interval, cond, soc, alpha0, max_k,
  methode = "expo", abstol, reltol, minimum_multiple_births)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
