% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{approximate_brts}
\alias{approximate_brts}
\title{Approximate the branching times to a fixed resolution. Events separated by
 a time smaller than 10^-brts_precision can be considered simultaneous}
\usage{
approximate_brts(brts, brts_precision = 8)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{brts_precision}{set the level of approximation on the branching times.}
}
\description{
Approximate the branching times to a fixed resolution. Events separated by
 a time smaller than 10^-brts_precision can be considered simultaneous
}
