% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{BD.infer.lambda.from.mutations}
\alias{BD.infer.lambda.from.mutations}
\title{Does something}
\usage{
BD.infer.lambda.from.mutations(Nsubs, MBD.lambda, mu, age, N0 = 2,
  sequence_length = 1000, mutation_rate = 1/age, Nsteps = 40)
}
\description{
Does something
}
