% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{BD.Nmutations}
\alias{BD.Nmutations}
\title{Does something C}
\usage{
BD.Nmutations(lambda, mu, age, N0 = 2, sequence_length = 1000,
  mutation_rate = 1/age)
}
\arguments{
\item{age}{The age of the tree.}
}
\description{
Does something C
}
