% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{branchLengths}
\alias{branchLengths}
\title{Function returning branch lengths named by ending node}
\usage{
branchLengths(tr)
}
\arguments{
\item{tr}{a phylogeny of class \code{phylo}}
}
\description{
Function returning branch lengths named by ending node
}
