% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{brts2time_intervals_and_births}
\alias{brts2time_intervals_and_births}
\title{Converts branching times to "time intervals between branching times" and "birth at nodes" vectors}
\usage{
brts2time_intervals_and_births(brts)
}
\arguments{
\item{brts}{something}
}
\description{
Converts branching times to "time intervals between branching times" and "birth at nodes" vectors
}
