% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_tips_to_pars.R
\name{calc_diff_percentile}
\alias{calc_diff_percentile}
\title{Calculate the difference in cumulative probabilities for a given
  percentile}
\usage{
calc_diff_percentile(pars, n_0 = 2, tips = 100, lx = 2 * tips,
  age = 10, percentile = 0.99, methode = "lsodes", abstol = 1e-16,
  reltol = 1e-10)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{tips}{The number of tips that you don't want to exceed when you
simulate a tree (with a percent confidence of "percentile").}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{age}{the age of the tree.}

\item{percentile}{The area under the curve (aka probability mass) of the
distribution up to the given number of tips.}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
}}

\item{abstol}{absolute error tolerance for
the numerical integration using deSolve.}

\item{reltol}{relative error tolerance for
the numerical integration using deSolve.}
}
\value{
The parameters
}
\description{
Given the number of tips, it returns the difference between the
  cumulative probability to have less than that number of tips, and the
  required cumulative probability.
}
\author{
Giovanni Laudanno
}
