% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{calculate_conditional_probability1}
\alias{calculate_conditional_probability1}
\title{Internal mbd function}
\usage{
calculate_conditional_probability1(brts, pars, soc = 2,
  tips_interval = c(0, Inf), methode = "expo", abstol = 1e-16,
  reltol = 1e-10)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability.
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually 
do with a standard usage of mbd_sim.}

\item{methode}{Specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm} 
  \item \code{expo}: use \code{expoRkit} 
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{abstol}{something}

\item{reltol}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
