% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{calculate_conditional_probability2}
\alias{calculate_conditional_probability2}
\title{Internal MBD function}
\usage{
calculate_conditional_probability2(brts, pars, missing_tips_interval = c(0,
  Inf), soc = 2, alpha, methode = "expo", abstol = 1e-16,
  reltol = 1e-10, minimum_multiple_births = 0)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
