% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{calculate_condprob}
\alias{calculate_condprob}
\title{Calculate conditional probability: user mode}
\usage{
calculate_condprob(pars, brts, lx, eq = mbd::condprob_select_eq(pars =
  pars, brts = brts, fortran = fortran), fortran = TRUE)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{eq}{the equation approach you want to use. It can be "q_eq" for the
Q-equation of "p_eq" for the P-equation.}

\item{fortran}{Set it to TRUE if you want to use FORTRAN routine.}
}
\description{
Calculate conditional probability: user mode
}
\author{
Giovanni Laudanno
}
