% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{calculate_condprob_nee_approx}
\alias{calculate_condprob_nee_approx}
\title{Calculate conditional probability using an approximation based on Nee et al.}
\usage{
calculate_condprob_nee_approx(pars, brts, n = 100)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{n}{maximum number of species that can speciate at the same time}
}
\description{
Calculate conditional probability using an approximation based on Nee et al.
}
\author{
Giovanni Laudanno
}
