% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{called_functions}
\alias{called_functions}
\title{Checks function calls within a function}
\usage{
called_functions(function.name, recursive = FALSE,
  checked.functions = NULL)
}
\description{
Checks function calls within a function
}
