% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{cat2}
\alias{cat2}
\title{cat2}
\usage{
cat2(message, verbose)
}
\arguments{
\item{verbose}{choose if you want to print the output or not}
}
\value{
prints on screen
}
\description{
If verbose == TRUE cats the message, otherwise stays silent
}
\author{
Giovanni Laudanno
}
