% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_brts}
\alias{check_brts}
\title{Check 'brts'}
\usage{
check_brts(brts, n_0)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}
}
\value{
nothing
}
\description{
Make sure that 'brts' are properly defined.
}
\author{
Giovanni Laudanno
}
