% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{check_brts_consistency}
\alias{check_brts_consistency}
\title{Test consistency of branching times}
\usage{
check_brts_consistency(brts, N0)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{N0}{the number of lineages at time equals zero.}
}
\value{
TRUE or FALSE.
}
\description{
Test if the given branching times can actually be generated by a MBD process. In any moment you CANNOT have more births than number of species.
}
\author{
Giovanni Laudanno
}
