% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_cond}
\alias{check_cond}
\title{Check 'cond'}
\usage{
check_cond(cond, tips_interval, n_0)
}
\arguments{
\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
}}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{n_0}{the number of lineages at time equals zero.}
}
\value{
nothing
}
\description{
Make sure that 'cond' and 'tips_interval' are coherent
}
\author{
Giovanni Laudanno
}
