% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_condprob_eq}
\alias{check_condprob_eq}
\title{Check "equation" for conditional probability}
\usage{
check_condprob_eq(eq)
}
\arguments{
\item{eq}{the equation approach you want to use. It can be "q_eq" for the
Q-equation of "p_eq" for the P-equation.}
}
\value{
Nothing
}
\description{
Check "equation" for conditional probability
}
\author{
Giovanni Laudanno
}
