% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_lx}
\alias{check_lx}
\title{Check lx}
\usage{
check_lx(lx)
}
\arguments{
\item{lx}{it is the number of ODEs considered for the computation.}
}
\value{
Nothing
}
\description{
Check lx
}
\author{
Giovanni Laudanno
}
