% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_n_0}
\alias{check_n_0}
\title{Check 'n_0'}
\usage{
check_n_0(n_0)
}
\arguments{
\item{n_0}{the number of lineages at time equals zero.}
}
\value{
nothing
}
\description{
Check 'n_0'
}
\author{
Giovanni Laudanno
}
