% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_pars}
\alias{check_pars}
\title{Check mbd pars}
\usage{
check_pars(pars, safety_checks = TRUE, lambda_limit = 10,
  mu_limit = 10, nu_limit = 100, q_threshold = 0.001)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{safety_checks}{TRUE if you want the parameters to be limited to a
realistic, but not infinite, parameter space}

\item{lambda_limit}{Upper limit to lambda estimations.}

\item{mu_limit}{Upper limit to mu estimations.}

\item{nu_limit}{Upper limit to nu estimations.}

\item{q_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}
}
\value{
Returns "wrong" if they are wrong, or "right" otherwise.
}
\description{
Check mbd pars
}
\author{
Giovanni Laudanno
}
