% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_q_vector}
\alias{check_q_vector}
\title{Check q_t}
\usage{
check_q_vector(q_t, t, pars, brts, debug_mode = FALSE)
}
\arguments{
\item{q_t}{the q_vector in time}

\item{t}{time}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}
}
\value{
Nothing
}
\description{
Check q_t
}
\author{
Giovanni Laudanno
}
