% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{check_sum_probs}
\alias{check_sum_probs}
\title{Check the vectors of probability sums computed during likelihood integration}
\usage{
check_sum_probs(sum_probs_1, sum_probs_2, debug_mode)
}
\arguments{
\item{sum_probs_1}{vector of probability sums}

\item{sum_probs_2}{vector of probability sums}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}
}
\description{
Check the vectors of probability sums computed during likelihood integration
}
