% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob}
\alias{cond_prob}
\title{Called by \link{mbd_loglik} if there is a conditioning != 0}
\usage{
cond_prob(pars, brts, cond, n_0 = 2, tips_interval = c(n_0 * (cond >
  0), Inf), lx = 30, debug_mode = FALSE)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}
}
\value{
the conditional probability
}
\description{
Called by \link{mbd_loglik} if there is a conditioning != 0
}
\author{
Giovanni Laudanno, Bart Haegeman
}
