% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_dp_nu}
\alias{cond_prob_dp_nu}
\title{Calculates the nu component of dp}
\usage{
cond_prob_dp_nu(pp, nu_matrix)
}
\arguments{
\item{pp}{the matrix p_{n1, n2}}

\item{nu_matrix}{matrix for the nu component. Component {m,n} is equal to
choose(n, m - n) x q ^ (m - n) x (1 - q) ^ (2 * n - m)}
}
\description{
Calculates the nu component of dp
}
\author{
Giovanni Laudanno, Bart Haegeman
}
