% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_dq_mu}
\alias{cond_prob_dq_mu}
\title{Calculates the mu component of dq}
\usage{
cond_prob_dq_mu(qq, qq2, k, m1, m2, mm)
}
\arguments{
\item{qq}{the matrix q_{m1, m2}}

\item{qq2}{the matrix q_{m1, m2}, with an empty frame}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{m1}{a matrix keeping track of columns}

\item{m2}{a matrix keeping track of rows}

\item{mm}{indexes of missing species}
}
\description{
Calculates the mu component of dq
}
