% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_p}
\alias{cond_prob_p}
\title{Called by \link{mbd_loglik} if there is a conditioning != 0}
\usage{
cond_prob_p(pars, brts, cond, n_0 = 2, lx = 30, tips_interval = c(n_0
  * (cond > 0), Inf), debug_mode = FALSE)
}
\value{
the conditional probability
}
\description{
Called by \link{mbd_loglik} if there is a conditioning != 0
}
\author{
Giovanni Laudanno, Bart Haegeman
}
