% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_p2_rhs1}
\alias{cond_prob_p2_rhs1}
\title{Auxilary function for cond_prob_p}
\usage{
cond_prob_p2_rhs1(pvec, lambda, mu, nu, nu_matrix, m1, m2, empty_pp, t)
}
\description{
Auxilary function for cond_prob_p
}
\author{
Giovanni Laudanno, Bart Haegeman
}
