% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_q_rhs1}
\alias{cond_prob_q_rhs1}
\title{Auxilary function for cond_prob_q}
\usage{
cond_prob_q_rhs1(qvec, lambda, mu, nu, k, nu_matrix, m1, m2, empty_qq)
}
\description{
Auxilary function for cond_prob_q
}
\author{
Giovanni Laudanno, Bart Haegeman
}
