% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_sim}
\alias{cond_prob_sim}
\title{Estimates conditional probability using simulations}
\usage{
cond_prob_sim(pars, brts, cond = 1, n_0 = 2, lx = 30,
  tips_interval = c(n_0 * (cond > 0), Inf), n_sims = 10000,
  saveit = TRUE)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{n_sims}{number of simulations}

\item{saveit}{do you want to save the results?}
}
\description{
Estimates conditional probability using simulations
}
\author{
Giovanni Laudanno
}
