% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_dp}
\alias{condprob_dp}
\title{dp total}
\usage{
condprob_dp(pvec, parmsvec)
}
\arguments{
\item{pvec}{a vector of probabilities P}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}
}
\description{
dp total
}
\author{
Giovanni Laudanno
}
