% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{condprob_dp_absorb_rhs}
\alias{condprob_dp_absorb_rhs}
\title{Auxilary function for cond_prob_p, computing rhs}
\usage{
condprob_dp_absorb_rhs(t, x, parms)
}
\arguments{
\item{t}{time}

\item{x}{something}

\item{parms}{some parameters to pass on the ode function}
}
\description{
Auxilary function for cond_prob_p, computing rhs
}
\author{
Giovanni Laudanno, Bart Haegeman
}
