% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{condprob_dp_lambda_absorb}
\alias{condprob_dp_lambda_absorb}
\title{Calculates the lambda component of dp}
\usage{
condprob_dp_lambda_absorb(pp, pp2, m1, m2, mm)
}
\arguments{
\item{pp}{the matrix p_{n1, n2}}

\item{pp2}{the matrix p_{n1, n2}, with an empty frame}

\item{m1}{a matrix keeping track of columns}

\item{m2}{a matrix keeping track of rows}

\item{mm}{indexes of missing species}
}
\description{
Calculates the lambda component of dp
}
\author{
Giovanni Laudanno, Bart Haegeman
}
