% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_dp_rhs}
\alias{condprob_dp_rhs}
\title{dp to be passed to \code{desolve::ode}}
\usage{
condprob_dp_rhs(t, x, parms)
}
\arguments{
\item{t}{time}

\item{x}{something}

\item{parms}{some parameters to pass on the ode function}
}
\description{
dp to be passed to \code{desolve::ode}
}
\author{
Giovanni Laudanno
}
