% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_dq_mu}
\alias{condprob_dq_mu}
\title{Calculates the mu component of dq}
\usage{
condprob_dq_mu(qq, qq2, m1, m2, mm)
}
\arguments{
\item{qq}{the matrix q_{m1, m2}}

\item{qq2}{the matrix q_{m1, m2}, with an empty frame}

\item{m1}{a matrix keeping track of columns}

\item{m2}{a matrix keeping track of rows}

\item{mm}{indexes of missing species}
}
\description{
Calculates the mu component of dq
}
