% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_empty_mat}
\alias{condprob_empty_mat}
\title{Creates empty matrix for condprob}
\usage{
condprob_empty_mat(lx_top)
}
\arguments{
\item{lx_top}{maximum lx to produce pre-calculated matrices for condprob}
}
\description{
Creates empty matrix for condprob
}
\author{
Giovanni Laudanno
}
