% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_log_q_mat}
\alias{condprob_log_q_mat}
\title{Creates log q matrix for the specified equation for condprob}
\usage{
condprob_log_q_mat(lx, q, eq)
}
\arguments{
\item{lx}{it is the number of ODEs considered for the computation.}

\item{q}{the single-lineage speciation probability at a triggered event.}

\item{eq}{the equation approach you want to use. It can be "q_eq" for the
Q-equation of "p_eq" for the P-equation.}
}
\description{
Creates log q matrix for the specified equation for condprob
}
\author{
Giovanni Laudanno
}
