% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_m1_mat}
\alias{condprob_m1_mat}
\title{Creates m1 matrix for condprob}
\usage{
condprob_m1_mat(lx_top)
}
\arguments{
\item{lx_top}{maximum lx to produce pre-calculated matrices for condprob}
}
\description{
Creates m1 matrix for condprob
}
\author{
Giovanni Laudanno
}
