% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_m2_mat}
\alias{condprob_m2_mat}
\title{Creates m2 matrix for condprob}
\usage{
condprob_m2_mat(m1_mat)
}
\arguments{
\item{m1_mat}{a matrix keeping track of columns}
}
\description{
Creates m2 matrix for condprob
}
\author{
Giovanni Laudanno
}
