% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{condprob_nee}
\alias{condprob_nee}
\title{Conditional probability using Nee et al's approach}
\usage{
condprob_nee(brts, parmsvec)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}
}
\description{
Conditional probability using Nee et al's approach
}
\author{
Giovanni Laudanno
}
