% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{condprob_nu_matrix_q}
\alias{condprob_nu_matrix_q}
\title{Creates vector of parameters for the conditional probability calculated with
the P-equation}
\usage{
condprob_nu_matrix_q(pars, lx, absorb)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{absorb}{set absorbing states when maximum number of allowed species is
reached}
}
\value{
nu_matrix for the P equation
}
\description{
Creates vector of parameters for the conditional probability calculated with
the P-equation
}
\author{
Giovanni Laudanno
}
