% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_peq_log_nu_mat}
\alias{condprob_peq_log_nu_mat}
\title{Creates log nu matrix for the P-equation for condprob}
\usage{
condprob_peq_log_nu_mat(lx)
}
\arguments{
\item{lx}{it is the number of ODEs considered for the computation.}
}
\description{
Creates log nu matrix for the P-equation for condprob
}
\author{
Giovanni Laudanno
}
