% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_peq_log_q_mat}
\alias{condprob_peq_log_q_mat}
\title{Creates log q matrix for the P-equation for condprob}
\usage{
condprob_peq_log_q_mat(lx, q)
}
\arguments{
\item{lx}{it is the number of ODEs considered for the computation.}

\item{q}{the single-lineage speciation probability at a triggered event.}
}
\description{
Creates log q matrix for the P-equation for condprob
}
\author{
Giovanni Laudanno
}
