% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_q_m1_m2}
\alias{condprob_q_m1_m2}
\title{Conditional probability Q(m1, m2) for all the states m1 and m2}
\usage{
condprob_q_m1_m2(brts, parmsvec, lx, rhs_function = mbd::condprob_dq_rhs)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{rhs_function}{a function for the right hand side of the differential
equation}
}
\description{
Conditional probability Q(m1, m2) for all the states m1 and m2
}
\author{
Giovanni Laudanno
}
