% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{condprob_select_eq}
\alias{condprob_select_eq}
\title{Selects the best eq for condprob}
\usage{
condprob_select_eq(pars, fortran = TRUE)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{fortran}{Set it to TRUE if you want to use FORTRAN routine.}
}
\description{
Selects the best eq for condprob
}
\author{
Giovanni Laudanno
}
