% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{condprob_sim}
\alias{condprob_sim}
\title{Estimates conditional probability using simulations}
\usage{
condprob_sim(brts, parmsvec, lx = 10000, saveit = TRUE, starting_seed = 1)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{saveit}{do you want to save the results?}

\item{starting_seed}{the seed to start from}
}
\description{
Estimates conditional probability using simulations
}
\author{
Giovanni Laudanno
}
