% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{count_n_mb_events}
\alias{count_n_mb_events}
\title{Count the number of multiple speciation events}
\usage{
count_n_mb_events(brts)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}
}
\description{
Count the number of multiple speciation events
}
\author{
Giovanni Laudanno
}
