% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{count_percentage_mb_events}
\alias{count_percentage_mb_events}
\title{Count the percentage of multiple speciation events}
\usage{
count_percentage_mb_events(brts)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}
}
\description{
Count the percentage of multiple speciation events
}
\author{
Giovanni Laudanno
}
