% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_A.no_mbd}
\alias{create_A.no_mbd}
\title{Internal MBD function}
\usage{
create_A.no_mbd(lambda, mu, nu, q, k, max_number_of_species,
  minimum_multiple_births)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
