% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{create_a_cond_2}
\alias{create_a_cond_2}
\title{Transition matrix for cond = 2}
\usage{
create_a_cond_2(pars, k, lx)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{lx}{it is the number of ODEs considered for the computation.}
}
\description{
Creates the A matrix,
used for the cond = 2 calculation
}
\details{
This is not to be called by the user.
}
\author{
Giovanni Laudanno
}
