% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_a_zero}
\alias{create_a_zero}
\title{Internal mbd function}
\usage{
create_a_zero(max_number_of_species, lambda, mu, q, k,
  matrix_builder = hyper_a_hanno)
}
\arguments{
\item{max_number_of_species}{something}

\item{lambda}{something}

\item{mu}{something}

\item{q}{a speciation parameter}

\item{k}{something}

\item{matrix_builder}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
