% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_B0}
\alias{create_B0}
\title{Internal mbd function}
\usage{
create_B0(max_number_of_species, q, k, b, matrix_builder = hyperA_HannoX)
}
\arguments{
\item{max_number_of_species}{something}

\item{q}{something}

\item{k}{something}

\item{b}{something}

\item{matrix_builder}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
