% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_b_no_mbd}
\alias{create_b_no_mbd}
\title{Internal mbd function}
\usage{
create_b_no_mbd(lambda, nu, q, k, b, max_number_of_species,
  minimum_multiple_births)
}
\arguments{
\item{lambda}{the sympatric speciation rate}

\item{nu}{the multiple allopatric speciation trigger rate}

\item{q}{the single-lineage speciation probability at a triggered event}

\item{k}{something}

\item{b}{something}

\item{max_number_of_species}{something}

\item{minimum_multiple_births}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
