% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_main_utils.R
\name{create_data_file_name}
\alias{create_data_file_name}
\title{Create data file name}
\usage{
create_data_file_name(data_folder, sim_pars, optim_ids, cond, n_0, t_0s,
  seed)
}
\arguments{
\item{data_folder}{The data folder insider the project folder.}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{n_0}{the number of lineages at time equals zero.}

\item{t_0s}{starting time for each clade}

\item{seed}{the seed}
}
\description{
Create data file name
}
