% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params.R
\name{create_mbd_params}
\alias{create_mbd_params}
\title{Create the MBD parameters}
\usage{
create_mbd_params(lambda = lambda, mu = mu, nu = nu, q = q)
}
\description{
Create the MBD parameters
}
\author{
Richel J.C. Bilderbeek
}
