% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params_selector.R
\name{create_mbd_params_selector}
\alias{create_mbd_params_selector}
\title{Create an MBD parameter selector}
\usage{
create_mbd_params_selector(lambda = FALSE, mu = FALSE, nu = FALSE,
  q = FALSE)
}
\description{
Create an MBD parameter selector
}
\author{
Richel J.C. Bilderbeek
}
