% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{default_lx}
\alias{default_lx}
\title{Default value for lx}
\usage{
default_lx(brts, missnumspec = 0)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}
}
\description{
Default value for lx
}
