% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{deliver_loglik}
\alias{deliver_loglik}
\title{Delivers the likelihood at the end of \link{mbd_loglik}}
\usage{
deliver_loglik(likelihood, sum_probs_1, sum_probs_2, cond, pc, debug_mode)
}
\arguments{
\item{likelihood}{the likelihood}

\item{sum_probs_1}{vector of probability sums, used during the
integration of \link{mbd_loglik}}

\item{sum_probs_2}{vector of probability sums, used during the
integration of \link{mbd_loglik}}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
}}

\item{pc}{conditional probability}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}
}
\description{
Delivers the likelihood at the end of \link{mbd_loglik}
}
