% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{determine_k_limit}
\alias{determine_k_limit}
\title{Internal mbd function}
\usage{
determine_k_limit(pars, brts, lx, soc, methode, abstol = 1e-16,
  reltol = 1e-10)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability.
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{lx}{It is the number of ODEs considered for the computation.}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{methode}{Specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm} 
  \item \code{expo}: use \code{expoRkit} 
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{abstol}{something}

\item{reltol}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
