% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{estimate_n_taxa}
\alias{estimate_n_taxa}
\title{Estimate approximately the expected number of taxa}
\usage{
estimate_n_taxa(pars, n_0, age)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{age}{the age of the tree.}
}
\value{
a list with mean and standard deviation
}
\description{
Estimate approximately the expected number of taxa
}
