% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_sim_utils.R
\name{evaluate_sim}
\alias{evaluate_sim}
\title{Establish whether the outcome of the simulation is in accordance with the
 required conditioning}
\usage{
evaluate_sim(l_matrix, cond, n_0, tips_interval, pool)
}
\arguments{
\item{l_matrix}{the l-table}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{pool}{the pool of living species at a given time}
}
\description{
Establish whether the outcome of the simulation is in accordance with the
 required conditioning
}
\author{
Giovanni Laudanno
}
