% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Does something}
\usage{
extract_posterior(file_name = "simcophylo_1_RUN1.(time).trees.txt",
  maxtree = 10)
}
\arguments{
\item{file_name}{file name}

\item{maxtree}{something}
}
\description{
Does something
}
