% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{find_best_lx_for_Pc}
\alias{find_best_lx_for_Pc}
\title{Internal MBD function}
\usage{
find_best_lx_for_Pc(brts, pars, soc = 2, methode = "expo",
  abstol = 1e-16, reltol = 1e-10, iterations = 20,
  interval.min = 500, interval.max = 1400)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
