% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_tips_to_pars.R
\name{find_nu_q_relation}
\alias{find_nu_q_relation}
\title{Find nu-q relation}
\usage{
find_nu_q_relation(tips = 100, age = 10, start_pars = c(0.2, 0.15, 1,
  0.15), chain_length = 40, verbose = FALSE)
}
\arguments{
\item{tips}{The number of tips that you don't want to exceed when you
simulate a tree 99 times out of 100.}

\item{age}{the age of the tree.}

\item{start_pars}{starting parameter values for the ml process.}

\item{chain_length}{The number of different combination of starting
parameters.}

\item{verbose}{choose if you want to print the output or not}
}
\value{
The parameters
}
\description{
Find the exponent of the formula nu*q^x that minimizes the
 differences across all runs in the chain.
}
\author{
Giovanni Laudanno
}
