% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{get_branch_lengths}
\alias{get_branch_lengths}
\title{Function returning branch lengths named by ending node}
\usage{
get_branch_lengths(tr)
}
\arguments{
\item{tr}{a phylogeny of class \code{phylo}}
}
\description{
Function returning branch lengths named by ending node
}
