% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_collect_cluster_data.R
\name{get_data_from_cluster}
\alias{get_data_from_cluster}
\title{Does something G}
\usage{
get_data_from_cluster(sim_pars, s, account_name = "p274829")
}
\arguments{
\item{sim_pars}{something}

\item{s}{something}

\item{account_name}{something}
}
\description{
Does something G
}
