% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{get_full_filename}
\alias{get_full_filename}
\title{Create an empty phylogeny}
\usage{
get_full_filename(pars, age)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{age}{the age of the tree.}
}
\description{
Create an empty phylogeny
}
\author{
Giovanni Laudanno
}
