% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_sim_utils.R
\name{get_info_l_matrix}
\alias{get_info_l_matrix}
\title{Extract information from the l-table}
\usage{
get_info_l_matrix(l_matrix, brts_precision)
}
\arguments{
\item{l_matrix}{the l-table}

\item{brts_precision}{set the level of approximation on the branching times.}
}
\description{
Extract information from the l-table
}
\author{
Giovanni Laudanno
}
