% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{get_lx_condprob}
\alias{get_lx_condprob}
\title{Fix the value for lx for condprob in \link{mbd_loglik}}
\usage{
get_lx_condprob(lx)
}
\arguments{
\item{lx}{it is the number of ODEs considered for the computation.}
}
\description{
Fix the value for lx for condprob in \link{mbd_loglik}
}
