% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_condprob.R
\name{get_lx_condprob_slow}
\alias{get_lx_condprob_slow}
\title{Fix the value for lx for condprob in \link{mbd_loglik}}
\usage{
get_lx_condprob_slow(pars, n_0, age, lx)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{age}{the age of the tree.}

\item{lx}{it is the number of ODEs considered for the computation.}
}
\description{
Fix the value for lx for condprob in \link{mbd_loglik}
}
