% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mbd_param_names.R
\name{get_mbd_param_names}
\alias{get_mbd_param_names}
\title{Get the names of the parameters used in the MBD model}
\usage{
get_mbd_param_names()
}
\description{
Get the names of the parameters used in the MBD model
}
\author{
Richel J.C. Bilderbeek
}
