% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{get_model_names}
\alias{get_model_names}
\title{Check if provided models make sense}
\usage{
get_model_names(function_names, verbose = FALSE)
}
\arguments{
\item{verbose}{choose if you want to print the output or not}
}
\value{
models names
}
\description{
Check if provided models make sense
}
\author{
Giovanni Laudanno
}
