% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{get_nee_pars}
\alias{get_nee_pars}
\title{Calculate the approximate Nee et al's equivalent parameters}
\usage{
get_nee_pars(pars)
}
\description{
Calculate the approximate Nee et al's equivalent parameters
}
