% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{hyper_a_hanno}
\alias{hyper_a_hanno}
\title{Internal mbd function}
\usage{
hyper_a_hanno(n_species, k, q)
}
\arguments{
\item{n_species}{number of species}

\item{k}{something}

\item{q}{a speciation parameter}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
