% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{hyper_matrix1}
\alias{hyper_matrix1}
\title{Function to build a matrix, used in creating the A and B operators.
It produces the structure
 q ^ (m - n) * (1 - q) ^ (k + 2 * n-m) *
 sum_j 2 ^ j choose(k, j) * choose(n, m - n - j)}
\usage{
hyper_matrix1(n_species, k, q)
}
\arguments{
\item{n_species}{number of species.}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{q}{the single-lineage speciation probability at a triggered event.}
}
\description{
Function to build a matrix, used in creating the A and B operators.
It produces the structure
 q ^ (m - n) * (1 - q) ^ (k + 2 * n-m) *
 sum_j 2 ^ j choose(k, j) * choose(n, m - n - j)
}
\author{
Hanno Hildenbrandt
}
