% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{hyperA_HannoX}
\alias{hyperA_HannoX}
\title{Internal mbd function}
\usage{
hyperA_HannoX(n_species, k, q)
}
\arguments{
\item{n_species}{number of species}

\item{k}{something}

\item{q}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
