% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_sim_utils.R
\name{initialize_l_matrix}
\alias{initialize_l_matrix}
\title{Initialize the l-table}
\usage{
initialize_l_matrix(age, n_0)
}
\arguments{
\item{age}{the age of the tree.}

\item{n_0}{the number of lineages at time equals zero.}
}
\description{
Initialize the l-table
}
\author{
Giovanni Laudanno
}
