% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{initialize_q_t}
\alias{initialize_q_t}
\title{Initialize the vector q_t for \link{mbd_loglik}}
\usage{
initialize_q_t(lx, lt)
}
\arguments{
\item{lx}{it is the number of ODEs considered for the computation.}

\item{lt}{length of time intervals vector}
}
\description{
Initialize the vector q_t for \link{mbd_loglik}
}
