% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mbd_params.R
\name{is_mbd_params}
\alias{is_mbd_params}
\title{Is object \code{x} an MBD parameter set,
  as can be created by \code{create_mbd_params}?}
\usage{
is_mbd_params(x)
}
\arguments{
\item{x}{the object to be determined of if
its an MBD parameter set}
}
\value{
TRUE if yes, else FALSE
}
\description{
Is object \code{x} an MBD parameter set,
  as can be created by \code{create_mbd_params}?
}
\author{
Richel J.C. Bilderbeek
}
