% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mbd_params_selector.R
\name{is_mbd_params_selector}
\alias{is_mbd_params_selector}
\title{Determine if the supplied object in an MBD parameter selector,
as can be created by \code{create_mbd_params_selector}}
\usage{
is_mbd_params_selector(x)
}
\arguments{
\item{x}{the object to be determined if it is an MBD
parameters selector}
}
\value{
TRUE if x is an MBD parameters selector
}
\description{
Determine if the supplied object in an MBD parameter selector,
as can be created by \code{create_mbd_params_selector}
}
\examples{
  s <- create_mbd_params_selector()
  testthat::expect_true(is_mbd_params_selector(s))

  testthat::expect_false(is_mbd_params_selector("nonsense"))
}
\author{
Richel J.C. Bilderbeek
}
