% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{is_pbd}
\alias{is_pbd}
\title{Test for pure birth in \link{mbd_loglik}}
\usage{
is_pbd(pars, tips_interval, cond, missnumspec, n_0)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}
}
\description{
Test for pure birth in \link{mbd_loglik}
}
