% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{matrix_check}
\alias{matrix_check}
\title{Checks if some matrix entries are infinite, NaN, NA
  or (only in the lower triangle) negative}
\usage{
matrix_check(m_list, sign_check = 0)
}
\arguments{
\item{m_list}{something}

\item{sign_check}{something}
}
\description{
Checks if some matrix entries are infinite, NaN, NA
  or (only in the lower triangle) negative
}
