% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{mbd_count_n_spec_events}
\alias{mbd_count_n_spec_events}
\title{Count the number of multiple speciation events}
\usage{
mbd_count_n_spec_events(brts)
}
\description{
Count the number of multiple speciation events
}
\author{
Giovanni Laudanno
}
