% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{mbd_experiment_logliks}
\alias{mbd_experiment_logliks}
\title{Logliks for the experiment}
\usage{
mbd_experiment_logliks()
}
\value{
loglik functions to use for the experiment
}
\description{
Get the loglik functions to use for the experiment
}
\author{
Giovanni Laudanno
}
