% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{mbd_experiment_pars}
\alias{mbd_experiment_pars}
\title{Generates parameters for the mbd experiment}
\usage{
mbd_experiment_pars(lambdas = c(0.2), mus = c(0, 0.1), nus = c(0,
  0.5, 1, 1.5), qs = c(0.1, 0.15, 0.2), cond = 1, age = 8)
}
\arguments{
\item{lambdas}{the sympatric speciation rates.}

\item{mus}{the extinction rates.}

\item{nus}{the multiple allopatric speciation trigger rates.}

\item{qs}{the single-lineage speciation probabilities at a triggered event.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
}}

\item{age}{the age of the tree.}
}
\value{
the parameter table for the experiment
}
\description{
Generates parameters for the mbd experiment
}
\author{
Giovanni Laudanno
}
