% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_integrate.R
\name{mbd_integrate}
\alias{mbd_integrate}
\title{performs integration of the ode}
\usage{
mbd_integrate(y, times, func, parms, atol, rtol, tcrit)
}
\arguments{
\item{y}{something}

\item{times}{vector of times}

\item{func}{a function}

\item{parms}{some parameters to pass on the ode function}

\item{atol}{absolute error tolerance for
the numerical integration using deSolve.}

\item{rtol}{relative error tolerance for
the numerical integration using deSolve.}

\item{tcrit}{the time you don't want to exceed in the integration}
}
\description{
performs integration of the ode
}
\author{
Rampal S. Etienne
}
