% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik.R
\name{mbd_loglik}
\alias{mbd_loglik}
\title{Calculates the likelihood for a multiple birth-death process}
\usage{
mbd_loglik(pars, brts, n_0 = 2, cond = 1, missnumspec = 0,
  lx = min(1 + 2 * (length(brts) + max(missnumspec)), mbd::max_lx()),
  tips_interval = c(n_0 * (cond > 0), Inf), q_threshold = 0.001,
  debug_mode = FALSE, rhs_function = "mbd_runmod")
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{q_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}

\item{rhs_function}{a function for the right hand side of the differential
equation}
}
\value{
The function returns the natural logarithm
  of the likelihood for the process.
}
\description{
mbd_loglik provides the likelihood for a process
  in which multiple births (from different parents) at the same time
  are possible, along with usual sympatric speciation and extinction events.
}
\author{
Giovanni Laudanno
}
