% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik.R
\name{mbd_loglik0}
\alias{mbd_loglik0}
\title{Calculates the likelihood for a multiple birth-death process}
\usage{
mbd_loglik0(pars, brts, soc = 2, cond = 0, tips_interval = c(0, Inf),
  missnumspec = 0, safety_threshold = 1e-04, methode = "expo",
  debug_check = 0, alpha = 20)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability.
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{soc}{Sets whether stem or crown age should be used (1 or 2)}

\item{cond}{Set 1 if you want to condition on stem or crown age and non-extinction of the phylogeny. Set 0 otherwise.}

\item{tips_interval}{It takes into account tips boundaries constrain on simulated dataset.}

\item{missnumspec}{The number of species that are in the clade but missing in the phylogeny.}

\item{safety_threshold}{It determines the precision on the parameters.}

\item{methode}{Specifies how the integration must be performed: set "sexpm" if you want to use sexpm; set "expo" if you want to use expoRkit; set "lsoda" if you want to use the "lsoda" method with the "deSolve::ode" function.}
}
\value{
The function returns the natural logarithm of the likelihood for the process.
}
\description{
mbd_loglik0 provides the likelihood for a process in which multiple births (from different parents) at the same time are possible.
}
\examples{
set.seed(11)
simulated_data = MBD:::mbd_sim0( pars=c(2.5,0.1,0.1),soc=2,age=10,cond=1 )
plot(simulated_data$tas)
mbd_loglik0( pars=c(1.2,0.05,0.1),brts=simulated_data$brts,soc=2,cond=1,missnumspec=0 )

}
\author{
Giovanni Laudanno
}
