% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_choosepar.R
\name{mbd_loglik_choosepar}
\alias{mbd_loglik_choosepar}
\title{Transformed loglik function to plug into the mle routine}
\usage{
mbd_loglik_choosepar(trparsopt, trparsfix, idparsopt = 1:4,
  idparsfix = (1:4)[-idparsopt], brts, n_0 = 2, cond = 1,
  tips_interval = c(0, Inf), missnumspec = 0, lx = 1 + 2 *
  (length(brts) + length(missnumspec)), methode = "expo")
}
\arguments{
\item{trparsopt}{something}

\item{trparsfix}{something}

\item{idparsopt}{the ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{idparsfix}{something}

\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}
}
\description{
This function provides a likelihood for a subset of parameters.
This is built to work inside mbd_ml.
}
\details{
This is not to be called by the user.
}
