% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_choosepar.R
\name{mbd_loglik_choosepar0}
\alias{mbd_loglik_choosepar0}
\title{Internal MBD function}
\usage{
mbd_loglik_choosepar0(trparsopt, trparsfix, idparsopt = 1:3,
  idparsfix = (1:3)[-idparsopt], brts, cond = 1, soc = 2,
  tips_interval = c(0, Inf), missnumspec = 0, methode = "expo",
  alpha = 20, pars.transform = 0)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
