% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{mbd_logliks_function}
\alias{mbd_logliks_function}
\title{Loglik functions in mbd package}
\usage{
mbd_logliks_function()
}
\value{
loglik functions
}
\description{
Get all the loglik functions in mbd package
}
\author{
Giovanni Laudanno
}
