% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{mbd_methodes}
\alias{mbd_methodes}
\title{Integration methodes}
\usage{
mbd_methodes()
}
\value{
the integration methodes
}
\description{
Gives the integration methodes (methods for odes)
 accepted by mbd
}
\author{
Giovanni Laudanno
}
