% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ml.R
\name{mbd_ml}
\alias{mbd_ml}
\title{Maximization of the loglikelihood
  under a multiple birth-death diversification model}
\usage{
mbd_ml(start_pars = c(0.3, 0.15, 2, 0.1), true_pars,
  optim_ids = c(TRUE, TRUE, TRUE, TRUE), brts, n_0 = 2, cond = 1,
  tips_interval = c(0, Inf), missnumspec = 0, lx = 1 + 2 *
  (length(brts) + length(missnumspec)), optimmethod = "simplex",
  methode = "expo", verbose = TRUE)
}
\arguments{
\item{start_pars}{starting parameter values for the ml process.}

\item{true_pars}{true parameter values when running the ml process.}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{optimmethod}{optimization routine: choose between subplex and simplex}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{verbose}{choose if you want to print the output or not}
}
\value{
The output is a dataframe containing
  estimated parameters and maximum loglikelihood.
  The computed loglikelihood contains the factor q! m! / (q + m)!
  where q is the number of species in the phylogeny and m is the number of
  missing species, as explained in the supplementary
  material to Etienne et al. 2012.
}
\description{
mbd_ml computes the maximum likelihood estimates of
  the parameters of a multiple birth-death diversification model
  for a given set of phylogenetic branching times.
  It also outputs the corresponding loglikelihood
  that can be used in model comparisons.
  Differently from mbd_ml it can account for three kind of events:
  sympatric (single) speciation, multiple (allopatric) speciation
  and extinction.
}
\examples{
set.seed(2)
lambda <- 0.2 # sympatric speciation rate
mu <- 0.15 # extinction rate;
nu <- 2.0 # multiple allopatric speciation trigger rate
q <- 0.1 # single-lineage speciation probability
sim_pars <- c(lambda, mu, nu, q)
crown_age <- 1
cond <- 1
n_0 <- 2
sim <- mbd_sim(
 pars = sim_pars,
 n_0 = n_0, # Use a crown age
 age = crown_age,
 cond = cond # Condition on non-extinction
)
start_pars <- c(0.2, 0.15, 1, 0.15)
optim_ids <- c(FALSE, FALSE, FALSE, TRUE)
graphics::plot(simulated_data$reconstructed_tree)
out <- mbd::mbd_ml(
  start_pars = start_pars,
  true_pars = sim_pars,
  optim_ids = optim_ids,
  brts = out$brts,
  cond = cond,
  n_0 = n_0,
  verbose = TRUE
)
}
\seealso{
use \link{pmb_ml} to perform a maximum likelihood estimation
  for a Pure Multiple Birth model.
}
\author{
Giovanni Laudanno
}
