% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ML.R
\name{mbd_ML0}
\alias{mbd_ML0}
\title{Maximization of the loglikelihood under a multiple birth-death diversification model}
\usage{
mbd_ML0(brts, initparsopt, idparsopt, idparsfix = (1:3)[-idparsopt], parsfix,
  missnumspec = 0, cond = 1, soc = 2, tips_interval = c(0, Inf),
  res = 10 * (1 + length(brts) + missnumspec), tol = c(0.001, 1e-04, 1e-06),
  maxiter = 1000 * round((1.25)^length(idparsopt)),
  changeloglikifnoconv = FALSE, optimmethod = "subplex", methode = "expo",
  alpha = 20, pars.transform = 1)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{initparsopt}{The initial values of the parameters that must be optimized}

\item{idparsopt}{The ids of the parameters that must be optimized. The ids are defined as follows:
\itemize{
\item id == 1 corresponds to lambda (multiple speciation trigger rate)
\item id == 2 corresponds to mu (extinction rate)
\item id == 3 corresponds to q (single-lineage speciation probability)
}}

\item{idparsfix}{The ids of the parameters that should not be optimized. The default is to fix all parameters not specified in idparsopt.}

\item{parsfix}{The values of the parameters that should not be optimized.}

\item{missnumspec}{The number of species that are in the clade but missing in the phylogeny.}

\item{cond}{Set 1 if you want to condition on stem or crown age and non-extinction of the phylogeny. Set 0 otherwise.}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{tips_interval}{It takes into account tips boundaries constrain on simulated dataset.}

\item{res}{Sets the maximum number of species for which a probability must be computed, must be larger than 1 + length(brts).}

\item{tol}{Sets the tolerances in the optimization. Consists of:
\itemize{
\item reltolx = relative tolerance of parameter values in optimization
\item reltolf = relative tolerance of function value in optimization
\item abstolx = absolute tolerance of parameter values in optimization
}}

\item{changeloglikifnoconv}{If TRUE the loglik will be set to -Inf if ML does not converge.}

\item{optimmethod}{Method used in optimization of the likelihood. Current default is 'subplex'. Alternative is 'simplex' (default of previous versions).}

\item{methode}{Set "sexpm" if you want to use sexpm. Set "expo" if you want to use expoRkit. Set "lsoda" if you want to use "lsoda".}

\item{max_iter}{Sets the maximum number of iterations in the optimization.}
}
\value{
The output is a dataframe containing estimated parameters and maximum
loglikelihood. The computed loglikelihood contains the factor q! m! / (q + m)!
where q is the number of species in the phylogeny and m is the number of
missing species, as explained in the supplementary material to Etienne et al. 2012.
}
\description{
mbd_ML0 computes the maximum likelihood estimates of the parameters of a multiple birth-death diversification model for a given set of phylogenetic branching times. It also outputs the corresponding loglikelihood that can be used in model comparisons.
}
\examples{
set.seed(11)
test_pars = c(1.6,0.1,0.08)
simulated_data = MBD:::mbd_sim0( pars=test_pars,soc=2,age=10,cond=1 )
plot(simulated_data$tas)
MBD:::mbd_ML0(brts=simulated_data$brts, initparsopt = 0.11 ,idparsopt = 3,
idparsfix = 1:2 ,parsfix = test_pars[1:2],missnumspec=0,cond=1, soc = 2)
}
\author{
Giovanni Laudanno
}
