% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ML.R
\name{mbd_ML_cluster}
\alias{mbd_ML_cluster}
\title{Maximization of the loglikelihood under a multiple birth-death diversification model, wrapped for cluster usage.}
\usage{
mbd_ML_cluster(s, initparsopt = c(0.6, 0.1, 1.3, 0.16))
}
\arguments{
\item{s}{The number of the simulation you want to evaluate.}

\item{initparsopt}{The initial values of the parameters that must be optimized}

\item{idparsopt}{The ids of the parameters that must be optimized. The ids are defined as follows:
\itemize{
\item id == 1 corresponds to lambda (speciation rate)
\item id == 2 corresponds to mu (extinction rate)
\item id == 3 corresponds to nu (multiple speciation trigger rate)
\item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{parsfix}{The values of the parameters that should not be optimized.}
}
\value{
The output is saved on the document "mbd_MLE.txt".
\itemize{
\item First column contains ML estimates for lambda.
\item Second column contains ML estimates for mu.
\item Third column contains ML estimates for lambda.
\item Fourth column contains maximum likelihood.
\item Fifth column contains the number of additional species coming from multiple births in the evaluated tree.
}
}
\description{
mbd_ML_cluster computes the maximum likelihood estimates of the parameters of a multiple birth-death diversification model. Differently from mbd_ML it needs only the number "s" of the simulations, making it suitable to run on a cluster. You will need two files to make it work: "general_settings","sim_data"; they are both generated by mbd_sim_dataset. This second version supports both sympatric and allopatric speciations.
}
\examples{
#You will need two files to make it work: "general_settings","sim_data".
MBD:::mbd_ML_cluster(1)

}
\author{
Giovanni Laudanno
}
