% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{mbd_p_eq}
\alias{mbd_p_eq}
\title{Provides estimated number of species integrating the P-equation}
\usage{
mbd_p_eq(test_parameters, age = 15, max_number_of_species = 2000,
  precision = 50L, soc = 2, output = 0)
}
\arguments{
\item{test_parameters}{something}

\item{age}{The age of the tree.}

\item{max_number_of_species}{something}

\item{precision}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{output}{something}
}
\description{
Provides estimated number of species integrating the P-equation
}
