% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_sim.R
\name{mbd_sim}
\alias{mbd_sim}
\title{Creates simulated trees under the multiple birth death process,
  including both sympatric and allopatric speciation}
\usage{
mbd_sim(pars, n_0 = 2, age = 10, cond = 1, seed = NA,
  tips_interval = c(0, Inf), brts_precision = 8)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{age}{the age of the tree.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{seed}{the seed}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{brts_precision}{set the level of approximation on the branching times.}
}
\description{
mbd_sim produces simulated trees allowing for
  three kind of events: sympatric speciation, multiple allopatric
  speciations and extinction.
}
\examples{
out <- mbd_sim(
  pars = c(0.6, 0.1, 0.4, 0.1),
  n_0 = 2,
  age = 10,
  cond = 1,
  tips_interval = c(0, Inf)
)
graphics::plot(out$full_tree)
graphics::plot(out$reconstructed_tree)
out$l_matrix

}
\author{
Giovanni Laudanno
}
