% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_sim.R
\name{mbd_sim_dataset0}
\alias{mbd_sim_dataset0}
\title{Creates a full simulated dataset of trees under the multiple birth death process}
\usage{
mbd_sim_dataset0(sim_pars = c(2.5, 0.1, 0.1), soc = 2, cond = 1,
  age = 10, max_sims = 1000, tips_interval = c(0, Inf), edge = Inf,
  minimum_multiple_births = 0)
}
\arguments{
\item{soc}{stands for stem or crown. Set 1 for stem or 2 for crown.}

\item{cond}{conditions on the survival of the crown (or the stem). Set to 1 if you want to condition on crown survival. Else set 0.}

\item{age}{is the age of the tree.}

\item{tips_interval}{You can also define the tips_interval as you can usually do with a standard usage of mbd_sim0.}

\item{edge}{the program automatically detects the (estimated) average number of tips. "edge" defines the width of the spread around the mean.}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability.
}}
}
\value{
The function returns a list of brts vectors, one for each of the max_sims simulations.
It also saves those in a file called "sim_data" and saves all the settings in a "general_settings" file.
N.B.: At each call of the function you overwrite the previous files.
}
\description{
mbd_sim_dataset0 produces a full dataset of max_sims simulated trees including multiple speciations at the same time.
}
\examples{
# @Giappo: does not work
# out <- mbd_sim_dataset0( pars=c(2.5,0.1,0.1),soc=2,age=10,cond=1,edge=Inf )

}
\author{
Giovanni Laudanno
}
