% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{mbd_solve}
\alias{mbd_solve}
\title{mbd ODE system integrator}
\usage{
mbd_solve(q_vector, time_interval, func = mbd_loglik_rhs, matrix_a,
  debug_mode = FALSE)
}
\arguments{
\item{q_vector}{the \code{Q} vector from
'Etienne et al. 2012 - Proc. R. Soc. B'.}

\item{time_interval}{something}

\item{func}{function for the right hand side of the ODE}

\item{matrix_a}{The A matrix from the theory that represents the ODE system.}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}
}
\description{
Integrates "func" in the time interval
}
\author{
Hanno Hildenbrandt, adapted by Giovanni Laudanno
}
