% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_integrate.R
\name{mbd_solve_methode}
\alias{mbd_solve_methode}
\title{mbd ODE system integrator for a specific ode method}
\usage{
mbd_solve_methode(
  vector,
  time_interval,
  func = mbd::mbd_loglik_rhs,
  parms,
  methode = "lsoda"
)
}
\arguments{
\item{vector}{a vector}

\item{time_interval}{a time interval}

\item{func}{function for the right hand side of the ODE}

\item{parms}{some parameters to pass on the ode function}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
  \item \code{ode45}: use \code{ode45} and \code{deSolve::ode}
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}
}
\description{
Integrates "func" in the time interval
}
\author{
Hanno Hildenbrandt, adapted by Giovanni Laudanno
}
