% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_tips_to_pars.R
\name{mbd_tips_to_pars}
\alias{mbd_tips_to_pars}
\title{Returns best pars to have the process producing
  less than a given number of tips, with a given probability.}
\usage{
mbd_tips_to_pars(tips = 100, age = 10, n_0 = 2, percentile = 0.99,
  start_pars = c(0.2, 0.15, 1.5, 0.15), optim_ids = c(FALSE, FALSE,
  TRUE, TRUE), chain_length = 10^sum(optim_ids), methode = "ode45",
  verbose = TRUE)
}
\arguments{
\item{tips}{The number of tips that you don't want to exceed when you
simulate a tree (with a percent confidence of "percentile").}

\item{age}{the age of the tree.}

\item{n_0}{the number of lineages at time equals zero.}

\item{percentile}{The area under the curve (aka probability mass) of the
distribution up to the given number of tips.}

\item{start_pars}{starting parameter values for the ml process.}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{chain_length}{The number of different combination of starting
parameters.}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
}}

\item{verbose}{choose if you want to print the output or not}
}
\value{
The parameters
}
\description{
Same as "mbd_tips_to_pars_single", but it returns many possible
  initial conditions. It is useful in case of more than one parameter.
}
\author{
Giovanni Laudanno
}
