% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_tips_to_pars.R
\name{mbd_tips_to_pars_single}
\alias{mbd_tips_to_pars_single}
\title{Find the best parameters to have a fixed probability to simulate
  trees with less than the given number of tips}
\usage{
mbd_tips_to_pars_single(start_pars = c(0.2, 0.15, 1.5, 0.15),
  tips = 100, percentile = 0.99, age = 10, n_0 = 2,
  optim_ids = c(FALSE, FALSE, TRUE, TRUE), methode = "ode45",
  verbose = TRUE)
}
\arguments{
\item{start_pars}{starting parameter values for the ml process.}

\item{tips}{The number of tips that you don't want to exceed when you
simulate a tree (with a percent confidence of "percentile").}

\item{percentile}{The area under the curve (aka probability mass) of the
distribution up to the given number of tips.}

\item{age}{the age of the tree.}

\item{n_0}{the number of lineages at time equals zero.}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
}}

\item{verbose}{choose if you want to print the output or not}
}
\value{
The parameters
}
\description{
It returns the parameters that minimize the difference between
  the requested percentile and the cumulative probability for the given
  number of tips.
}
\author{
Giovanni Laudanno
}
