% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{myheatmap}
\alias{myheatmap}
\title{Does something
Gets heatmap of my matrices}
\usage{
myheatmap(matrix, logs = 1, colormap = heat.colors(15), ...)
}
\description{
Does something
Gets heatmap of my matrices
}
