% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{nee_stdev_nt}
\alias{nee_stdev_nt}
\title{Average number of species for b-d process.
See Kendall 1948a, pag. 12}
\usage{
nee_stdev_nt(nee_pars, n_0, age)
}
\arguments{
\item{nee_pars}{the equivalent parameters in Nee et al.'s approach}

\item{n_0}{the number of lineages at time equals zero.}

\item{age}{the age of the tree.}
}
\description{
Average number of species for b-d process.
See Kendall 1948a, pag. 12
}
