% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{negatives_correction}
\alias{negatives_correction}
\title{Checks for NA, NaN or negative components in a vector (usually used for Qt)}
\usage{
negatives_correction(v, pars, display_output = 0)
}
\arguments{
\item{v}{sometjing}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability_
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{display_output}{something}
}
\description{
Checks for NA, NaN or negative components in a vector (usually used for Qt)
}
