% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{percentiles_function}
\alias{percentiles_function}
\title{Percentiles function}
\usage{
percentiles_function(results, sim_pars, printit = 1,
  quantiles_choice = c(0.25, 0.5, 0.75))
}
\arguments{
\item{sim_pars}{something}
}
\description{
Percentiles function
}
