% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{prob_cond_get_p_m1_m2}
\alias{prob_cond_get_p_m1_m2}
\title{Solutiof the integration of the ODE for p_{n1, n2}}
\usage{
prob_cond_get_p_m1_m2(pars, brts, matrices,
  rhs_function = cond_prob_p_rhs2)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{matrices}{matrices}

\item{rhs_function}{a function for the right hand side of the differential
equation}
}
\description{
Solutiof the integration of the ODE for p_{n1, n2}
}
\author{
Giovanni Laudanno, Bart Haegeman
}
