% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{summarize_beast_posterior}
\alias{summarize_beast_posterior}
\title{Does something}
\usage{
summarize_beast_posterior(input_trees_path, input_xml = NULL,
  subsamp = NULL)
}
\arguments{
\item{input_trees_path}{Path to \code{.trees} file}

\item{input_xml}{something}

\item{subsamp}{something}
}
\description{
Does something
}
