% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{A_operator}
\alias{A_operator}
\title{Internal MBD function}
\usage{
A_operator(Q, transition_matrix, time_interval, precision = 50L,
  A_abstol = 1e-16, A_reltol = 1e-10, methode = "expo")
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
