% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{A_operator}
\alias{A_operator}
\title{Internal mbd function}
\usage{
A_operator(Q, transition_matrix, time_interval, precision = 50L,
  A_abstol = 1e-16, A_reltol = 1e-10, methode = "expo")
}
\arguments{
\item{Q}{something}

\item{transition_matrix}{something}

\item{time_interval}{something}

\item{precision}{something}

\item{A_abstol}{something}

\item{A_reltol}{something}

\item{methode}{Specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm} 
  \item \code{expo}: use \code{expoRkit} 
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
