% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{alpha_analysis}
\alias{alpha_analysis}
\title{Internal MBD function}
\usage{
alpha_analysis(brts, pars, tips_interval, cond, soc, alpha0, max_k,
  methode = "expo", abstol, reltol, minimum_multiple_births)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.}

\item{cond}{Set 1 if you want to condition on stem or crown age and non-extinction of the phylogeny. Set 0 otherwise.}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
