% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{alpha_analysis}
\alias{alpha_analysis}
\title{Internal mbd function}
\usage{
alpha_analysis(brts, pars, tips_interval, cond, soc, alpha0, max_k,
  methode = "expo", abstol, reltol, minimum_multiple_births)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability_
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{cond}{Set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{soc}{Sets whether stem or crown age should be used. Use '1' for stem,
use '2' for crown.}

\item{alpha0}{something}

\item{max_k}{something}

\item{methode}{Specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{abstol}{something}

\item{reltol}{something}

\item{minimum_multiple_births}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
