% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{are_these_parameters_wrong}
\alias{are_these_parameters_wrong}
\title{are_these_parameters_wrong}
\usage{
are_these_parameters_wrong(brts, pars, safety_threshold, n_0)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{n_0}{the number of lineages at time equals zero.}
}
\description{
check parameters consistency
}
\author{
Giovanni Laudanno
}
