% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{brts2time_intervals_and_births}
\alias{brts2time_intervals_and_births}
\title{Converts branching times to 'time intervals between branching times'
  and 'birth at nodes' vectors}
\usage{
brts2time_intervals_and_births(brts, brts_precision = 8)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{brts_precision}{set the level of approximation on the branching times.}
}
\description{
Converts branching times to 'time intervals between branching times'
  and 'birth at nodes' vectors
}
