% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{calculate_conditional_probability}
\alias{calculate_conditional_probability}
\title{Internal MBD function}
\usage{
calculate_conditional_probability(brts, pars, soc = 2, tips_interval = c(0,
  Inf), methode = "expo", abstol = 1e-16, reltol = 1e-10)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
