% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{calculate_conditional_probability}
\alias{calculate_conditional_probability}
\title{Internal MBD function}
\usage{
calculate_conditional_probability(brts, pars, lx = 1000, soc = 2,
  tips_interval = c(0, Inf), methode = "expo", abstol = 1e-16,
  reltol = 1e-10)
}
\arguments{
\item{brts}{something}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{lx}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.}

\item{methode}{something}

\item{abstol}{something}

\item{reltol}{something}
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
