% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{called_functions}
\alias{called_functions}
\title{Checks function calls within a function}
\usage{
called_functions(function_name, checked_functions = NULL)
}
\arguments{
\item{function_name}{something}

\item{checked_functions}{something}
}
\description{
Checks function calls within a function
}
