% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_checks.R
\name{check_pc}
\alias{check_pc}
\title{Check conditional probability}
\usage{
check_pc(pc)
}
\arguments{
\item{pc}{conditional probability}
}
\value{
Nothing
}
\description{
Check conditional probability
}
\author{
Giovanni Laudanno
}
