% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{check_sum_probs}
\alias{check_sum_probs}
\title{Check the vectors of probability sums computed during likelihood integration}
\usage{
check_sum_probs(sum_probs_1, sum_probs_2)
}
\arguments{
\item{sum_probs_1}{vector of probability sums, used during the
integration of \link{mbd_loglik}}

\item{sum_probs_2}{vector of probability sums, used during the
integration of \link{mbd_loglik}}
}
\description{
Check the vectors of probability sums computed during likelihood integration
}
