% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{compare_functions}
\alias{compare_functions}
\title{Compares two functions that accept same arguments args=(...)}
\usage{
compare_functions(function1, function2, iterations = 1, output = 1,
  ...)
}
\arguments{
\item{function1}{something}

\item{function2}{something}

\item{iterations}{something}

\item{output}{something}

\item{...}{something}
}
\description{
Compares two functions that accept same arguments args=(...)
}
