% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_p_matrices}
\alias{cond_prob_p_matrices}
\title{Auxilary function for cond_prob_p}
\usage{
cond_prob_p_matrices(q, lx)
}
\description{
Auxilary function for cond_prob_p
}
\author{
Giovanni Laudanno, Bart Haegeman
}
