% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_conditioning.R
\name{cond_prob_q_rhs1}
\alias{cond_prob_q_rhs1}
\title{Auxilary function for cond_prob_q, creating rhs}
\usage{
cond_prob_q_rhs1(qvec, lambda, mu, nu, k, nu_matrix, m1, m2, empty_qq, t)
}
\arguments{
\item{qvec}{a vector of probabilities Q}

\item{lambda}{the sympatric speciation rate.}

\item{mu}{the extinction rate.}

\item{nu}{the multiple allopatric speciation trigger rate.}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{nu_matrix}{matrix for the nu component. Component {m,n} is equal to
choose(n, m - n) x q ^ (m - n) x (1 - q) ^ (2 * n - m)}

\item{m1}{a matrix keeping track of columns}

\item{m2}{a matrix keeping track of rows}

\item{empty_qq}{empty qq matrix. See \link{qq}.}

\item{t}{time}
}
\description{
Auxilary function for cond_prob_q, creating rhs
}
\author{
Giovanni Laudanno, Bart Haegeman
}
