% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob}
\alias{condprob}
\title{Conditional probability}
\usage{
condprob(brts, parmsvec, fortran = TRUE, eq)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}

\item{fortran}{Set it to TRUE if you want to use FORTRAN routine.}

\item{eq}{the equation approach you want to use. It can be "q_eq" for the
Q-equation of "p_eq" for the P-equation.}
}
\description{
Conditional probability
}
\author{
Giovanni Laudanno
}
