% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_dq}
\alias{condprob_dq}
\title{dq total}
\usage{
condprob_dq(qvec, parmsvec)
}
\arguments{
\item{qvec}{a vector of probabilities Q}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}
}
\description{
dq total
}
\author{
Giovanni Laudanno
}
