% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_fortran.R
\name{condprob_q}
\alias{condprob_q}
\title{Conditional probability calculated with the Q-approach}
\usage{
condprob_q(brts, parmsvec, fortran = TRUE, lx)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}

\item{fortran}{Set it to TRUE if you want to use FORTRAN routine.}

\item{lx}{it is the number of ODEs considered for the computation.}
}
\description{
Conditional probability calculated with the Q-approach
}
\author{
Giovanni Laudanno
}
