% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Imports data
No globals please}
\usage{
correlation_analysis(results, path, titolo = NULL, pdfname,
  sim_pars = sim_pars, percentage_hidden_outliers = 0.04, openit = 0,
  idparsopt, mother_folder)
}
\arguments{
\item{results}{something}

\item{path}{something}

\item{titolo}{something}

\item{pdfname}{something}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{percentage_hidden_outliers}{something}

\item{openit}{something}

\item{idparsopt}{The ids of the parameters that must be optimized. The ids are defined as follows:
\itemize{
\item id == 1 corresponds to lambda (speciation rate)
\item id == 2 corresponds to mu (extinction rate)
\item id == 3 corresponds to nu (multiple speciation trigger rate)
\item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{mother_folder}{something}

\item{parnames}{sonething
analysis function}
}
\description{
Imports data
No globals please
}
