% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_a}
\alias{create_a}
\title{A matrix}
\usage{
create_a(pars, k, lx, matrix_builder = mbd:::hyper_a_hanno)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{lx}{It is the number of ODEs considered for the computation.}

\item{matrix_builder}{function used to build the transition matrix. Default option is \code{hyper_a_hanno}}
}
\description{
Creates the A matrix, used for likelihood integration between branching times.
}
\details{
This is not to be called by the user.
}
\author{
Giovanni Laudanno
}
