% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_a}
\alias{create_a}
\title{The A matrix}
\usage{
create_a(pars, k, lx, matrix_builder = mbd::hyper_matrix2,
  no_species_out_of_the_matrix = FALSE)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{matrix_builder}{function used to build the transition matrix.
Default option is \code{hyper_matrix2}}

\item{no_species_out_of_the_matrix}{If true prevents interactions from states
describing number of species greater than lx}
}
\description{
Creates the A matrix,
used for likelihood integration between branching times.
}
\details{
This is not to be called by the user.
}
\author{
Giovanni Laudanno
}
