% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_b}
\alias{create_b}
\title{Internal mbd function}
\usage{
create_b(lambda, nu, q, k, b, max_number_of_species)
}
\arguments{
\item{lambda}{the sympatric speciation rate}

\item{nu}{the multiple allopatric speciation trigger rate}

\item{q}{the single-lineage speciation probability at a triggered event}

\item{k}{something}

\item{b}{something}

\item{max_number_of_species}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
