% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{create_B}
\alias{create_B}
\title{Internal MBD function}
\usage{
create_B(lambda, nu, q, k, b, max_number_of_species)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
