% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params_selector.R
\name{create_mbd_params_selector}
\alias{create_mbd_params_selector}
\title{Create an MBD parameter selector}
\usage{
create_mbd_params_selector(lambda = FALSE, mu = FALSE, nu = FALSE,
  q = FALSE)
}
\arguments{
\item{lambda}{the sympatric speciation rate}

\item{mu}{the extinction rate}

\item{nu}{the multiple allopatric speciation trigger rate}

\item{q}{the single-lineage speciation probability at a triggered event}
}
\description{
Create an MBD parameter selector
}
\author{
Richel J.C. Bilderbeek
}
