% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(a_abstol, a_reltol, abstol, account_name, after, age,
  alpha, alpha0, b, brts, chain_length, changeloglikifnoconv,
  checked_functions, colormap, cond, function_name, idparsfix, idparsopt,
  initparsopt, input_trees_path, input_xml, interval_max, interval_min,
  iterations, k, lambda, logs, lx, lx0, matrix, matrix_builder, max_iter,
  max_k, max_number_of_species, max_repetitions, max_sims, maxiter,
  mbd_lambda, methode, minimum_multiple_births, missnumspec, mu,
  mutation_rate, N, N0, n_steps, n_subs, nu, optimmethod, pars,
  pars_transform, parsfix, precision, print_errors, printit, Q, q,
  quantiles_choice, recursive, reltol, res, results, s, sample_interval,
  sequence_length, sim_pars, sim_phylo, soc, subsamp, t, t1, t2,
  tips_interval, time_interval, tol, tr, transition_matrix, trparsfix,
  trparsopt, values, verbose, x, x_name, x_splits, y, y_name, y_splits, z,
  z.name)
}
\arguments{
\item{a_abstol}{something}

\item{a_reltol}{something}

\item{abstol}{something}

\item{account_name}{something}

\item{after}{something}

\item{age}{The age of the tree.}

\item{alpha}{something}

\item{alpha0}{something}

\item{b}{something}

\item{brts}{A set of branching times of a phylogeny_}

\item{chain_length}{something}

\item{changeloglikifnoconv}{something}

\item{checked_functions}{something}

\item{colormap}{something}

\item{cond}{Set 1 if you want to condition on stem or crown age 
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{function_name}{something}

\item{idparsfix}{something}

\item{idparsopt}{The ids of the parameters that must be optimized. 
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{initparsopt}{something}

\item{input_trees_path}{Path to \code{.trees} file}

\item{input_xml}{something}

\item{interval_max}{something}

\item{interval_min}{something}

\item{iterations}{something}

\item{k}{something}

\item{lambda}{something}

\item{logs}{something}

\item{lx}{It is the number of ODEs considered for the computation.}

\item{lx0}{something}

\item{matrix}{something}

\item{matrix_builder}{something}

\item{max_iter}{Sets the maximum number of iterations in the optimization}

\item{max_k}{something}

\item{max_number_of_species}{something}

\item{max_repetitions}{something}

\item{max_sims}{something}

\item{maxiter}{something}

\item{mbd_lambda}{something}

\item{methode}{Specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm} 
  \item \code{expo}: use \code{expoRkit} 
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{minimum_multiple_births}{something}

\item{missnumspec}{The number of species that are in the clade but missing in the phylogeny_}

\item{mu}{something}

\item{mutation_rate}{something}

\item{N}{something}

\item{N0}{something}

\item{n_steps}{something}

\item{n_subs}{something}

\item{nu}{something}

\item{optimmethod}{something}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability_
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{pars_transform}{something}

\item{parsfix}{The values of the parameters that should not be optimized.}

\item{precision}{something}

\item{print_errors}{something}

\item{printit}{something}

\item{Q}{something}

\item{q}{something}

\item{quantiles_choice}{something}

\item{recursive}{something}

\item{reltol}{something}

\item{res}{something}

\item{results}{something}

\item{s}{something}

\item{sample_interval}{something}

\item{sequence_length}{something}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{sim_phylo}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{subsamp}{something}

\item{t}{something}

\item{t1}{something}

\item{t2}{something}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually 
do with a standard usage of mbd_sim.}

\item{time_interval}{something}

\item{tol}{something}

\item{tr}{a phylogeny of class \code{phylo}}

\item{transition_matrix}{something}

\item{trparsfix}{something}

\item{trparsopt}{something}

\item{values}{something}

\item{verbose}{something}

\item{x}{something}

\item{x_name}{something}

\item{x_splits}{something}

\item{y}{something}

\item{y_name}{something}

\item{y_splits}{something}

\item{z}{something}

\item{z.name}{something}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno, 
  use of this function by Richel J.C. Bilderbeek
}
