% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(a_abstol, a_reltol, abstol, account_name, after, age,
  alpha, alpha0, b, brts, brts_precision, changeloglikifnoconv,
  checked_functions, colormap, cond, function_name, idparsfix, idparsopt,
  initparsopt, input_trees_path, input_xml, interval_max, interval_min,
  iterations, k, lambda, loglik_function, logs, lx, lx0, matrix,
  matrix_builder, max_iter, max_k, max_number_of_species, max_repetitions,
  max_sims, maxiter, mbd_lambda, message, methode, minimum_multiple_births,
  missnumspec, mu, mutation_rate, n_0, n_species, init_n_lineages, n_steps,
  n_subs, nu, optimmethod, params, pars, pars_transform, pars_transformed,
  parsfix, precision, print_errors, printit, q_vector, q, quantiles_choice,
  recursive, reltol, res, results, s, seed, safety_threshold,
  sample_interval, sequence_length, sim_pars, sim_phylo, subsamp, t, t1,
  t2, tips_interval, time_interval, tol, tr, transition_matrix, trparsfix,
  trparsopt, values, verbose, x, x_name, x_splits, y, y_name, y_splits, z,
  z_name, start_pars, true_pars, optim_ids)
}
\arguments{
\item{a_abstol}{something}

\item{a_reltol}{something}

\item{abstol}{absolute error tolerance for
the numerical integration using deSolve.}

\item{account_name}{something}

\item{after}{something}

\item{age}{the age of the tree.}

\item{alpha}{something}

\item{alpha0}{something}

\item{b}{the number of simultaneous births on a given branching time.}

\item{brts}{A set of branching times of a phylogeny.}

\item{brts_precision}{set the level of approximation on the branching times.}

\item{changeloglikifnoconv}{something}

\item{checked_functions}{something}

\item{colormap}{something}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{function_name}{something}

\item{idparsfix}{something}

\item{idparsopt}{the ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{initparsopt}{something}

\item{input_trees_path}{Path to \code{.trees} file.}

\item{input_xml}{something}

\item{interval_max}{something}

\item{interval_min}{something}

\item{iterations}{something}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{lambda}{the sympatric speciation rate.}

\item{loglik_function}{the loglik function}

\item{logs}{something}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{lx0}{something}

\item{matrix}{something}

\item{matrix_builder}{function used to build the transition matrix.
Default option is \code{hyper_a_hanno}}

\item{max_iter}{Sets the maximum number of iterations in the optimization}

\item{max_k}{something}

\item{max_number_of_species}{something}

\item{max_repetitions}{something}

\item{max_sims}{something}

\item{maxiter}{something}

\item{mbd_lambda}{something}

\item{message}{a message}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{sexpm}: use \code{sexpm}
  \item \code{expo}: use \code{expoRkit}
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
}}

\item{minimum_multiple_births}{minimum amount of multiple births
that have to be present in the simulated phylogeny.}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{mu}{the extinction rate.}

\item{mutation_rate}{something}

\item{n_0}{the number of lineages at time equals zero.}

\item{n_species}{number of species.}

\item{init_n_lineages}{the number of lineages at time equals zero.}

\item{n_steps}{something}

\item{n_subs}{something}

\item{nu}{the multiple allopatric speciation trigger rate.}

\item{optimmethod}{optimization routine: choose between subplex and simplex}

\item{params}{transition matrix for the rhs of the ode system.}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{pars_transform}{something}

\item{pars_transformed}{parameters of the likelihood functions, transformed
according to y = x / (1 + x)}

\item{parsfix}{the values of the parameters that should not be optimized.}

\item{precision}{something}

\item{print_errors}{something}

\item{printit}{something}

\item{q_vector}{the \code{Q} vector from
'Etienne et al. 2012 - Proc. R. Soc. B'.}

\item{q}{the single-lineage speciation probability at a triggered event.}

\item{quantiles_choice}{something}

\item{recursive}{something}

\item{reltol}{relative error tolerance for
the numerical integration using deSolve.}

\item{res}{something}

\item{results}{something}

\item{s}{the seed}

\item{seed}{the seed}

\item{safety_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}

\item{sample_interval}{something}

\item{sequence_length}{something}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{sim_phylo}{something}

\item{subsamp}{something}

\item{t}{something}

\item{t1}{something}

\item{t2}{something}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{time_interval}{something}

\item{tol}{something}

\item{tr}{a phylogeny of class \code{phylo}}

\item{transition_matrix}{something}

\item{trparsfix}{something}

\item{trparsopt}{something}

\item{values}{something}

\item{verbose}{choose if you want to print the output or not}

\item{x}{something}

\item{x_name}{something}

\item{x_splits}{something}

\item{y}{something}

\item{y_name}{something}

\item{y_splits}{something}

\item{z}{something}

\item{z_name}{something}

\item{start_pars}{starting parameter values for the ml process.}

\item{true_pars}{true parameter values when running the ml process.}

\item{optim_ids}{ids of the parameters you want to optimize.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
