% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{determine_k_limit}
\alias{determine_k_limit}
\title{Internal MBD function}
\usage{
determine_k_limit(pars, brts, lx, soc, methode, abstol = 1e-16,
  reltol = 1e-10)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
