% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{find_best_lx_for_Pc}
\alias{find_best_lx_for_Pc}
\title{Internal mbd function}
\usage{
find_best_lx_for_Pc(brts, pars, soc = 2, methode = "expo",
  abstol = 1e-16, reltol = 1e-10, iterations = 20,
  interval.min = 500, interval.max = 1400)
}
\arguments{
\item{brts}{something}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability.
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{methode}{something}

\item{abstol}{something}

\item{reltol}{something}

\item{iterations}{something}

\item{interval.min}{something}

\item{interval.max}{something}
}
\description{
Internal mbd function.
}
\details{
This is not to be called by the user.
}
