% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_collect_cluster_data.R
\name{get_all_data_from_cluster}
\alias{get_all_data_from_cluster}
\title{Does something}
\usage{
get_all_data_from_cluster(sim_pars, max_sims = 1000,
  account_name = "p274829")
}
\description{
Does something
}
