% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{heatmap2dataframe}
\alias{heatmap2dataframe}
\title{Converts a matrix in a dataframe that can be used with ggplot}
\usage{
heatmap2dataframe(x, y, Matrix, x_name = "x", y.name = "y",
  heatmap.name = "HeatMap")
}
\arguments{
\item{x}{something}

\item{y}{something}

\item{Matrix}{something}

\item{x_name}{something}

\item{y.name}{something}

\item{heatmap.name}{heatmap name}
}
\description{
Converts a matrix in a dataframe that can be used with ggplot
}
