% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{hyper_a_hanno}
\alias{hyper_a_hanno}
\title{Function to build a matrix, used in creating the A and B operators.
It produces the structure
 q ^ (m - n) * (1 - q) ^ (k + 2 * n-m) *
 sum_j 2 ^ j choose(k, j) * choose(n, m - n - j)}
\usage{
hyper_a_hanno(n_species, k, q)
}
\arguments{
\item{n_species}{number of species.}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{q}{the single-lineage speciation probability at a triggered event.}
}
\description{
Function to build a matrix, used in creating the A and B operators.
It produces the structure
 q ^ (m - n) * (1 - q) ^ (k + 2 * n-m) *
 sum_j 2 ^ j choose(k, j) * choose(n, m - n - j)
}
\examples{
  m <- mbd::hyper_a_hanno(n_species = 2, k = 2, q = 0.1)
  testthat::expect_equal(m[1, 1], 0.81)
  testthat::expect_equal(m[1, 2], 0.00)
  testthat::expect_equal(m[1, 3], 0.00)
  testthat::expect_equal(m[2, 1], 0.36)
  testthat::expect_equal(m[2, 2], 0.729)
  testthat::expect_equal(m[2, 3], 0.00)
  testthat::expect_equal(m[3, 1], 0.04)
  testthat::expect_equal(m[3, 2], 0.405)
  testthat::expect_equal(m[3, 3], 0.6561)
}
\author{
Hanno Hildenbrandt, adapted by Richel J.C. Bilderbeek
}
