% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_main_utils.R
\name{main_save_files}
\alias{main_save_files}
\title{Save data and results}
\usage{
main_save_files(
  project_folder,
  sim_pars,
  optim_ids,
  cond,
  n_0,
  t_0s,
  seed,
  sim,
  results
)
}
\arguments{
\item{project_folder}{the folder when you want to save data and results}

\item{sim_pars}{vector of starting parameters:
\itemize{
  \item sim_pars[1] is lambda, the sympatric speciation rate;
  \item sim_pars[2] is mu, the extinction rate;
  \item sim_pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item sim_pars[4] is q, the single-lineage speciation probability;
}}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{t_0s}{starting time for each clade}

\item{seed}{the seed}

\item{sim}{the results of a sim run}

\item{results}{mle results}
}
\description{
Save data and results
}
