% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_integrate.R
\name{mbd_integrate}
\alias{mbd_integrate}
\title{performs integration of the ode}
\usage{
mbd_integrate(y = y, times = tseq, func = func, parms = parms,
  atol = atol, rtol = rtol, tcrit = t1)
}
\arguments{
\item{y}{something}

\item{parms}{some parameters to pass on the ode function}
}
\description{
performs integration of the ode
}
\author{
Rampal S. Etienne
}
