% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_choosepar.R
\name{mbd_loglik_choosepar}
\alias{mbd_loglik_choosepar}
\title{Internal MBD function}
\usage{
mbd_loglik_choosepar(trparsopt, trparsfix, idparsopt = 1:4,
  idparsfix = (1:4)[-idparsopt], brts, cond = 1, soc = 2,
  tips_interval = c(0, Inf), missnumspec = 0, methode = "expo",
  minimum_multiple_births = 0, pars.transform = 0, print_errors = 0,
  ...)
}
\description{
Internal MBD function.
}
\details{
This is not to be called by the user.
}
