% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{mbd_loglik_rhs}
\alias{mbd_loglik_rhs}
\title{Builds the right hand side of the ODE set for multiple birth model}
\usage{
mbd_loglik_rhs(t, x, params)
}
\arguments{
\item{t}{time}

\item{x}{something}

\item{params}{transition matrix for the rhs of the ode system.}
}
\description{
Builds the right hand side of the ODE set
for multiple birth model
}
\details{
This is not to be called by the user.
}
\author{
Hanno Hildebrandt
}
