% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_main.R
\name{mbd_main}
\alias{mbd_main}
\title{Run the full mbd routine}
\usage{
mbd_main(seed, sim_pars, cond = 1, n_0 = 2, age = 10,
  tips_interval = c(0, Inf), start_pars = c(0.3, 0.1, 1, 0.1),
  models = mbd_loglik, verbose = FALSE)
}
\arguments{
\item{seed}{the seed}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{n_0}{the number of lineages at time equals zero.}

\item{age}{the age of the tree.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{start_pars}{starting parameter values for the ml process.}

\item{verbose}{choose if you want to print the output or not}
}
\description{
Simulate an mbd process and infer parameters maximizing the
likelihood
}
\details{
mle inference
}
