% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_main.R
\name{mbd_main}
\alias{mbd_main}
\title{Run the full mbd routine}
\usage{
mbd_main(sim_pars, cond = 1, n_0 = 2, age = 10,
  tips_interval = c(n_0 * (cond > 0), Inf), seed, start_pars = c(0.3,
  0.1, 1.5, 0.15), optim_ids = rep(TRUE, length(start_pars)),
  loglik_functions = mbd_loglik, project_folder = NULL,
  verbose = FALSE, safety_threshold = 1e-04)
}
\arguments{
\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{n_0}{the number of lineages at time equals zero.}

\item{age}{the age of the tree.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{seed}{the seed}

\item{start_pars}{starting parameter values for the ml process.}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{loglik_functions}{the loglik_functions you want to use}

\item{project_folder}{the folder when you want to save data and results}

\item{verbose}{choose if you want to print the output or not}

\item{safety_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}
}
\description{
Simulate an mbd process and infer parameters maximizing the
likelihood
}
\details{
mle inference
}
