% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ml.R
\name{mbd_ml}
\alias{mbd_ml}
\title{Maximization of the loglikelihood
  under a multiple birth-death diversification model}
\usage{
mbd_ml(loglik_function = mbd_loglik, brts, start_pars = c(0.5, 0.3,
  0.5, 0.3), n_0 = 2, cond = 1, optim_ids = rep(TRUE,
  length(start_pars)), true_pars = start_pars, tips_interval = c(0,
  Inf), q_threshold = 0.001, verbose = TRUE, missnumspec = 0,
  lx = min(1 + 3 * (length(brts) + max(missnumspec)), mbd::max_lx()),
  maxiter = 10000)
}
\arguments{
\item{loglik_function}{the loglik function}

\item{brts}{A set of branching times of a phylogeny.}

\item{start_pars}{starting parameter values for the ml process.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{true_pars}{true parameter values when running the ml process.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{q_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}

\item{verbose}{choose if you want to print the output or not}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{maxiter}{maximum number of subplex iterations}
}
\value{
The output is a dataframe containing
  estimated parameters and maximum loglikelihood.
  The computed loglikelihood contains the factor q! m! / (q + m)!
  where q is the number of species in the phylogeny and m is the number of
  missing species, as explained in the supplementary
  material to Etienne et al. 2012.
}
\description{
mbd_ml computes the maximum likelihood estimates of
  the parameters of a multiple birth-death diversification model
  for a given set of phylogenetic branching times.
  It also outputs the corresponding loglikelihood
  that can be used in model comparisons.
  Differently from mbd_ml it can account for three kind of events:
  sympatric (single) speciation, multiple (allopatric) speciation
  and extinction.
}
\examples{
set.seed(2)
lambda <- 0.2 # sympatric speciation rate
mu <- 0.15 # extinction rate;
nu <- 2.0 # multiple allopatric speciation trigger rate
q <- 0.1 # single-lineage speciation probability
sim_pars <- c(lambda, mu, nu, q)
crown_age <- 1
cond <- 1
n_0 <- 2
sim <- mbd_sim(
 pars = sim_pars,
 n_0 = n_0, # Use a crown age
 age = crown_age,
 cond = cond # Condition on non-extinction
)
start_pars <- c(0.2, 0.15, 2, 0.15)
optim_ids <- c(FALSE, FALSE, FALSE, TRUE)
graphics::plot(sim$reconstructed_tree)
out <- mbd::mbd_ml(
  start_pars = start_pars,
  optim_ids = optim_ids,
  brts = sim$brts,
  true_pars = sim_pars,
  cond = cond,
  n_0 = n_0,
  verbose = FALSE
)
}
\author{
Giovanni Laudanno
}
