% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ml.R
\name{mbd_ml}
\alias{mbd_ml}
\title{Maximization of the loglikelihood
  under a multiple birth-death diversification model}
\usage{
mbd_ml(loglik_function = mbd_loglik, brts, start_pars = c(0.5, 0.3,
  0.5, 0.3), n_0 = 2, cond = 1, tips_interval = c(0, Inf),
  safety_threshold = 0.001, verbose = TRUE, optim_ids = rep(TRUE,
  length(start_pars)), true_pars = start_pars)
}
\arguments{
\item{loglik_function}{the loglik function}

\item{brts}{A set of branching times of a phylogeny.}

\item{start_pars}{starting parameter values for the ml process.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny. Set 0 otherwise.}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
You can also define the tips_interval as you can usually
do with a standard usage of mbd_sim.}

\item{safety_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}

\item{verbose}{choose if you want to print the output or not}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{true_pars}{true parameter values when running the ml process.}
}
\value{
The output is a dataframe containing
  estimated parameters and maximum loglikelihood.
  The computed loglikelihood contains the factor q! m! / (q + m)!
  where q is the number of species in the phylogeny and m is the number of
  missing species, as explained in the supplementary
  material to Etienne et al. 2012.
}
\description{
mbd_ml computes the maximum likelihood estimates of
  the parameters of a multiple birth-death diversification model
  for a given set of phylogenetic branching times.
  It also outputs the corresponding loglikelihood
  that can be used in model comparisons.
  Differently from mbd_ml it can account for three kind of events:
  sympatric (single) speciation, multiple (allopatric) speciation
  and extinction.
}
\examples{
set.seed(2)
lambda <- 0.2 # sympatric speciation rate
mu <- 0.15 # extinction rate;
nu <- 2.0 # multiple allopatric speciation trigger rate
q <- 0.1 # single-lineage speciation probability
sim_pars <- c(lambda, mu, nu, q)
crown_age <- 1
cond <- 1
n_0 <- 2
sim <- mbd_sim(
 pars = sim_pars,
 n_0 = n_0, # Use a crown age
 age = crown_age,
 cond = cond # Condition on non-extinction
)
start_pars <- c(0.2, 0.15, 2, 0.15)
optim_ids <- c(FALSE, FALSE, FALSE, TRUE)
graphics::plot(sim$reconstructed_tree)
out <- mbd::mbd_ml(
  start_pars = start_pars,
  optim_ids = optim_ids,
  brts = sim$brts,
  cond = cond,
  n_0 = n_0,
  verbose = FALSE
)
}
\author{
Giovanni Laudanno
}
