% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ml.R
\name{mbd_ml_cluster}
\alias{mbd_ml_cluster}
\title{Maximization of the loglikelihood under a multiple birth-death
 diversification model, wrapped for cluster usage.}
\usage{
mbd_ml_cluster(s, initparsopt = c(0.6, 0.1, 1.3, 0.16))
}
\arguments{
\item{s}{The number of the simulation you want to evaluate.}

\item{initparsopt}{The initial values of the parameters that
must be optimized}
}
\value{
The output is saved on the document "mbd_MLE.txt".
\itemize{
\item First column contains ML estimates for lambda.
\item Second column contains ML estimates for mu.
\item Third column contains ML estimates for lambda.
\item Fourth column contains maximum likelihood.
\item Fifth column contains the number of additional species coming
from multiple births in the evaluated tree.
}
}
\description{
mbd_ml_cluster computes the maximum likelihood
estimates of the parameters of a multiple birth-death diversification model.
Differently from mbd_ml it needs only the number "s" of the simulations,
making it suitable to run on a cluster. You will need two files
to make it work: "general_settings","sim_data";
they are both generated by mbd_sim_dataset.
This second version supports both sympatric and allopatric speciations.
This function is like Ulysses' bow: you cannot use it
unless you are the owner of the package.
}
\examples{
#You will need two files to make it work: "general_settings","sim_data".
# mbd_ml_cluster(1)

}
\author{
Giovanni Laudanno
}
