% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_integrate.R
\name{mbd_solve}
\alias{mbd_solve}
\title{mbd ODE system integrator}
\usage{
mbd_solve(vector, time_interval, func = mbd::mbd_loglik_rhs, parms)
}
\arguments{
\item{vector}{a vector}

\item{time_interval}{a time interval}

\item{func}{function for the right hand side of the ODE}

\item{parms}{some parameters to pass on the ode function}
}
\description{
Integrates "func" in the time interval
}
\author{
Giovanni Laudanno, Hanno Hildenbrandt
}
