% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_loglik_utils.R
\name{mbd_solve}
\alias{mbd_solve}
\title{mbd ODE system integrator}
\usage{
mbd_solve(vector, time_interval, func = mbd_loglik_rhs, parms)
}
\arguments{
\item{time_interval}{something}

\item{func}{function for the right hand side of the ODE}
}
\description{
Integrates "func" in the time interval
}
\author{
Hanno Hildenbrandt, adapted by Giovanni Laudanno
}
