% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{myheatmap}
\alias{myheatmap}
\title{Gets heatmap of my matrices.}
\usage{
myheatmap(matrix, logs = 1, colormap = heat.colors(15), ...)
}
\arguments{
\item{matrix}{something}

\item{logs}{something}

\item{colormap}{something}

\item{...}{something}
}
\description{
Gets heatmap of my matrices.
}
