% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{myheatmap2}
\alias{myheatmap2}
\title{Does something J}
\usage{
myheatmap2(x, y, z, x_name, y_name, z.name, x_splits, y_splits)
}
\arguments{
\item{x}{something}

\item{y}{something}

\item{z}{something}

\item{x_name}{something}

\item{y_name}{something}

\item{z.name}{something}

\item{x_splits}{something}

\item{y_splits}{something}
}
\description{
Does something J
}
