% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{percentiles_function}
\alias{percentiles_function}
\title{Percentiles function}
\usage{
percentiles_function(results, sim_pars, printit = 1,
  quantiles_choice = c(0.25, 0.5, 0.75))
}
\arguments{
\item{results}{something}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{printit}{something}

\item{quantiles_choice}{something}
}
\description{
Percentiles function
}
