% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_pmb.R
\name{pmb_loglik}
\alias{pmb_loglik}
\title{Calculate the log-likelihood for a Pure Multiple Birth model}
\usage{
pmb_loglik(pars, brts, n_0 = 2)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}
}
\description{
Calculate the log-likelihood for a Pure Multiple Birth model
}
\seealso{
use \link{mbd_loglik} to calculate the log-likelihood for
  a Multiple Birth Death model.
}
\author{
Giovanni Laudanno
}
