% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_pure_multiple_birth.R
\name{pmb_loglik_choosepar}
\alias{pmb_loglik_choosepar}
\title{Does something M}
\usage{
pmb_loglik_choosepar(trparsopt, trparsfix = 0, idparsopt = c(1, 3, 4),
  idparsfix = (1:4)[-idparsopt], brts, soc = 2, pars_transform = 0)
}
\arguments{
\item{trparsopt}{something}

\item{trparsfix}{something}

\item{idparsopt}{The ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{idparsfix}{something}

\item{brts}{A set of branching times of a phylogeny_}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{pars_transform}{something}
}
\description{
Does something M
}
