% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_pure_multiple_birth.R
\name{pmb_loglik_choosepar}
\alias{pmb_loglik_choosepar}
\title{Does something M}
\usage{
pmb_loglik_choosepar(trparsopt, trparsfix = 0, idparsopt = c(1, 3, 4),
  idparsfix = (1:4)[-idparsopt], brts, soc = 2, pars.transform = 0)
}
\arguments{
\item{trparsopt}{something}

\item{trparsfix}{something}

\item{idparsopt}{The ids of the parameters that must be optimized. The ids are defined as follows:
\itemize{
\item id == 1 corresponds to lambda (speciation rate)
\item id == 2 corresponds to mu (extinction rate)
\item id == 3 corresponds to nu (multiple speciation trigger rate)
\item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{idparsfix}{something}

\item{brts}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{pars.transform}{something}
}
\description{
Does something M
}
