% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_pmb.R
\name{pmb_loglik_choosepar}
\alias{pmb_loglik_choosepar}
\title{Provides a likelihood for a subset of parameters for
a Pure Multiple Birth model.
This function is built to work inside mbd_minusLL_vs_single_parameter
or any optimizer like optim or simplex}
\usage{
pmb_loglik_choosepar(trparsopt, trparsfix = 0, idparsopt = c(1, 3, 4),
  idparsfix = (1:4)[-idparsopt], brts, soc = 2, pars_transform = 0)
}
\arguments{
\item{trparsopt}{something}

\item{trparsfix}{something}

\item{idparsopt}{The ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{idparsfix}{something}

\item{brts}{A set of branching times of a phylogeny}

\item{soc}{Sets whether stem or crown age should be used. Use '1' for stem,
use '2' for crown.}

\item{pars_transform}{something}
}
\description{
Provides a likelihood for a subset of parameters for
a Pure Multiple Birth model.
This function is built to work inside mbd_minusLL_vs_single_parameter
or any optimizer like optim or simplex
}
\author{
Giovanni Laudanno
}
