% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_pure_multiple_birth.R
\name{pmb_ML}
\alias{pmb_ML}
\title{Does something N}
\usage{
pmb_ML(brts, initparsopt, soc = 2, res = 10 * (1 + length(brts) +
  missnumspec), tol = c(0.001, 1e-04, 1e-06), maxiter = 1000 *
  round((1.25)^length(idparsopt)), changeloglikifnoconv = FALSE,
  optimmethod = "simplex", pars.transform = 1)
}
\arguments{
\item{brts}{something}

\item{initparsopt}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{res}{something}

\item{tol}{something}

\item{maxiter}{something}

\item{changeloglikifnoconv}{something}

\item{optimmethod}{something}

\item{pars.transform}{something}
}
\description{
Does something N
}
