% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_pmb.R
\name{pmb_ml}
\alias{pmb_ml}
\title{Perform a maximum likelihood estimation for a Pure Multiple Birth model.}
\usage{
pmb_ml(brts, initparsopt, idparsopt = c(1, 3, 4), soc = 2, res = 10 * (1 +
  length(brts) + missnumspec), tol = c(0.001, 1e-04, 1e-06), maxiter = 1000
  * round((1.25)^length(idparsopt)), changeloglikifnoconv = FALSE,
  optimmethod = "simplex", pars_transform = 1, missnumspec = 0)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny}

\item{initparsopt}{something}

\item{idparsopt}{The ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{soc}{Sets whether stem or crown age should be used. Use '1' for stem,
use '2' for crown.}

\item{res}{something}

\item{tol}{something}

\item{maxiter}{something}

\item{changeloglikifnoconv}{something}

\item{optimmethod}{something}

\item{pars_transform}{something}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny_}
}
\description{
Perform a maximum likelihood estimation for a Pure Multiple Birth model.
}
\seealso{
use \link{mbd_ml} to perform a maximum likelihood estimation
  for a Multiple Birth Death model.
}
\author{
Giovanni Laudanno
}
