% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_main_utils.R
\name{print_info}
\alias{print_info}
\title{Print information about simulated data}
\usage{
print_info(brts, n_0, cond, verbose)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny_ Set 0 otherwise.}

\item{verbose}{choose if you want to print the output or not}
}
\value{
nothing
}
\description{
Print information about simulated data
}
\author{
Giovanni Laudanno
}
