% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_main_utils.R
\name{print_info}
\alias{print_info}
\title{Print information about simulated data}
\usage{
print_info(brts, n_0, cond, verbose)
}
\arguments{
\item{brts}{A set of branching times of a phylogeny.}

\item{n_0}{the number of lineages at time equals zero.}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{verbose}{choose if you want to print the output or not}
}
\value{
nothing
}
\description{
Print information about simulated data
}
\author{
Giovanni Laudanno
}
