% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_utils.R
\name{summarize_beast_posterior}
\alias{summarize_beast_posterior}
\title{Does something}
\usage{
summarize_beast_posterior(INPUT.TREES, INPUT.XML = NULL,
  subsamp = NULL)
}
\arguments{
\item{INPUT.TREES}{Path to \code{.trees} file}

\item{INPUT.XML}{something}

\item{subsamp}{something}
}
\description{
Does something
}
